requirements.txt
# Flask and extensions
flask==3.0.2
flask-cors==4.0.0
flask-jwt-extended==4.6.0
flask-mongoengine==1.0.0
flask-sqlalchemy==3.1.1
python-dotenv==1.0.1
flask-restx==1.3.0

# Database drivers
psycopg2-binary==2.9.9
pymongo==4.6.2

# API and Schema validation
marshmallow==3.21.1
requests==2.31.0
gql==3.5.0
requests-toolbelt==1.0.0

# Security
bcrypt==4.1.2
python-jose==3.3.0
passlib[bcrypt]==1.7.4

# Development
python-dotenv==1.0.1 